shardwallet: unify "split" and "merge" into "reforge"
Instead of two separate "split" and "merge" operations, we now have a
single "reforge" operation that takes an arbitrary number of parents and
produces an arbitrary number of children. The `split` and `merge` API
endpoints are kept for convenience.

This saves storage space on merges, because we no longer have to write
a `Children` structure for each parent of the merged node. Instead, we
write just one `ForgingData` structure per forging event. Forging events
are keyed by the ID of the first child token, which is required to exist
anyway (empty merges are forbidden), and which we need to fetch anyway
to compute the child index.

For now, `ForgingData` stores the `childrenSharesMicros`, which was
previously stored in `Children`. We could reduce storage here in two
ways. The first option is to compress down to a `uint24[]` (losslessly,
since micros values can't exceed 10^6, which is smaller than 2^24). Then
`computeClaimed` can read one storage slot per 10 siblings, which is
basically just a single read in practice. The second option is to not
store this at all, since the sibling shares can be recovered from
`shareMicros_`---but then we need to store the number of siblings
somewhere. If we limit a Shardwallet to 2^64 shards, we could pack into
one storage slot each token's share, first sibling, and number of
siblings. This would require `computeClaimed` to do multiple reads from
storage to read each sibling's share, though. Some profiling will be in
order.