shardwallet: in tests, only send tx from Hardhat signers
Sending transactions from a `Wallet` seems to break when running under
`solidity-coverage`, because some interception layer that causes the gas
price to be set to a small number does not affect `Wallet`s and so
causes an error that "sender doesn't have enough funds to send tx". The
correct value for the gas price differs between test contexts (i.e.,
with and without coverage). Seems like the easiest workaround is to just
not send transactions from `Wallet`s at all, which is unfortunate but
workable.