mintpass: implement auction mechanics
This patch starts a new contract for the QQL mint passes and auction.
The contract supports a Dutch auction of all the mint passes with a
common clearing price; buyers who purchased a mint pass above the
eventual clearing price can claim the difference as a rebate, even while
the auction is live. The price schedule structure is pretty flexible,
but might change later.

(This is a new contract rather than an update to `Ticket.sol` because
the goals and design have changed enough that that contract is not super
useful as a starting point. It also has some reentrancy bugs that this
new contract doesn't.)