shardwallet: add a view accessor for `getClaimAmount`
This requires some contortion for a parallel version of `computeClaimed`
that doesn't write to storage, and some more contortion to coax Solidity
into letting us call it. We opt for the "`staticcall`-to-self" approach,
rather than a function pointer cast, since that *should* be entirely
safe but requires inline assembly and isn't explicitly documented as
kosher.