shardwallet: return amounts claimed from `claimTo`
With this change, an `eth_call` RPC to `claim` will return the amount
that would actually be claimed. This provides the same functionality as
the `claimableAmount` function, but (a) is more efficient in the worst
case due to memoizing writes, (b) actually executes the same code as a
real claim, and (c) has a much more straightforward implementation.

This change also makes it easier for contracts that want to
programmatically interact with a shardwallet.