mintpass: include token IDs in purchase/reservation events
This avoids needing to correlate with `Transfer` events.

As a side-effect, this patch permutes the purchase/reservation events
after the actual transfers, which also means that the `firstTokenId`s on
events may not be in order (if the recipient of a pass has an ERC-721
receive hook that purchases more passes). This is something to be aware
of, but should be mostly surmountable: such permutations will only ever
occur within a single transaction, so any log parsing can simply sort
relevant events by `firstTokenId` within each transaction or block.