mintpass: add a raceless pause method
This just calls `updateAuctionSchedule` with a schedule that prevents
the current price from dropping further. This could be implemented as an
on-chain helper contract (it's not a strictly new primitive), but doing
so would require the owner to delegatecall through that helper, and
doing it manually is pretty prone to races across each price drop
boundary.