ERC721OperatorFilter: always allow mints and burns
Can't call `ownerOf` when the token doesn't exist yet. We work around
this by always allowing transfers when either the "from" argument or the
"to" argument is `null`, indicating a mint or a burn, respectively.
Much like always allowing users to transfer their own tokens, this seems
like something that's reasonable to always allow. And it fixes this bug.